+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - a11y: improvements to custom plugins modal (from #2677) (@rfiorentino1)
 - add bridge filter dropdown to accessories page
 - a11y: improvements to config restore modal (from #2677) (@rfiorentino1)
+- a11y: improvements to terminal component (from #2677) (@rfiorentino1)
 
 ### Other Changes